requirements.txt
# Core Dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Resume Parsing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.5.0